refactor(#255): extract dispatch.rs (dispatch, dispatchers, run_raw_passthrough)
Move prepend, prepend_without, extract_subcmd, passthrough_subcmd,
run_raw_passthrough, dispatch_cargo, dispatch_go, dispatch_swift,
dispatch_dotnet, all print_*_help functions, and the top-level
dispatch() router from mod.rs into a new crates/rskim/src/cmd/dispatch.rs.

Re-export dispatch and run_raw_passthrough from mod.rs so all callers
continue to work without modification. extract_subcmd, prepend_without,
and dispatch() sync-guard tests moved to dispatch.rs::tests.
Removed now-unused ExitCode import from mod.rs.

mod.rs drops from 1,032 to 494 lines. All 3,640+ tests pass.
